Update multi_actuator_transmission and delete the double_actuator_transmission
