+++ rm_hw/rm_hw_loader_plugins.xml
@@ -1,10 +1,10 @@
 <library path="lib/librm_hw_loader_plugins">
     <!-- Transmission types -->
-    <class name="transmission_interface/DoubleActuatorTransmission"
-           type="transmission_interface::DoubleActuatorTransmissionLoader"
+    <class name="transmission_interface/MultiActuatorTransmission"
+           type="transmission_interface::MultiActuatorTransmissionLoader"
            base_class_type="transmission_interface::TransmissionLoader">
         <description>
-            Load from a URDF description the configuration of a joint using two actuator transmission.
+            Load from a URDF description the configuration of a joint using multiple actuator transmission.
         </description>
     </class>
